requirements.txt
langchain==0.0.252
openai
faiss-cpu
streamlit
streamlit-chat
